+++ moco_paper_result.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+class MocoPaperResult(ABC):
+    @abstractmethod
+    def generate_results(self):
+        pass
+
+    @abstractmethod
+    def report_results(self):
+        pass
